+++ src/ORM/Repository.php
@@ -10,14 +10,14 @@ class Repository
     /** @var \Parable\ORM\Model */
     protected $model;
 
-    /** @var bool */
-    protected $onlyCount = false;
-
     /** @var array */
     protected $orderBy = [];
 
     /** @var array */
     protected $limitOffset = [];
+
+    /** @var bool */
+    protected $onlyCount = false;
 
     /** @var bool */
     protected $returnOne = false;
@@ -37,7 +37,7 @@ class Repository
     {
         $query = \Parable\ORM\Query::createInstance();
         $query->setTableName($this->getModel()->getTableName());
-        $query->setTableKey($this->getModel()->getTableKey());
+
         if ($this->onlyCount) {
             $query->select(['count(*)']);
         }
@@ -50,6 +50,7 @@ class Repository
         if ($this->returnOne) {
             $query->limitOffset(1);
         }
+
         return $query;
     }
 
@@ -104,21 +105,14 @@ class Repository
      * @param string     $key
      * @param string     $comparator
      * @param mixed|null $value
-     * @param string     $andOr
      *
      * @return \Parable\ORM\Model[]|\Parable\ORM\Model
      * @throws \Parable\ORM\Exception
      */
-    public function getByCondition($key, $comparator, $value = null, $andOr = \Parable\ORM\Query\ConditionSet::SET_AND)
+    public function getByCondition($key, $comparator, $value = null)
     {
         $query = $this->createQuery();
-        if ($andOr === \Parable\ORM\Query\ConditionSet::SET_AND) {
-            $conditionSet = $query->buildAndSet([$key, $comparator, $value]);
-        } elseif ($andOr === \Parable\ORM\Query\ConditionSet::SET_OR) {
-            $conditionSet = $query->buildOrSet([$key, $comparator, $value]);
-        } else {
-            throw new \Parable\ORM\Exception('Invalid andOr type given.');
-        }
+        $conditionSet = $query->buildAndSet([$key, $comparator, $value]);
         return $this->getByConditionSet($conditionSet);
     }
 
@@ -187,6 +181,19 @@ class Repository
     }
 
     /**
+     * Set onlyCount to true or false.
+     *
+     * @param bool $value
+     *
+     * @return $this
+     */
+    public function setOnlyCount($value = true)
+    {
+        $this->onlyCount = (bool)$value;
+        return $this;
+    }
+
+    /**
      * Sets the repo to return only one (the first), the same as getById always does.
      *
      * @return $this
@@ -243,19 +250,6 @@ class Repository
     }
 
     /**
-     * Set onlyCount to true or false.
-     *
-     * @param bool $value
-     *
-     * @return $this
-     */
-    public function onlyCount($value = true)
-    {
-        $this->onlyCount = (bool)$value;
-        return $this;
-    }
-
-    /**
      * Build and return an AND condition set.
      *
      * @param \Parable\ORM\Query\Condition[] $conditions
@@ -303,6 +297,17 @@ class Repository
     }
 
     /**
+     * Resets everything but the query and model class to their default values.
+     */
+    public function reset()
+    {
+        $this->orderBy = [];
+        $this->limitOffset = [];
+        $this->setOnlyCount(false);
+        $this->returnAll();
+    }
+
+    /**
      * Create an instance of the repository class for given $modelName.
      *
      * @param string $modelName
@@ -314,13 +319,20 @@ class Repository
         if (!class_exists($modelName)) {
             throw new Exception("Model '{$modelName}' does not exist.");
         }
+        return self::createForModel($modelName::create());
+    }
 
-        $model = $modelName::create();
-
+    /**
+     * Create an instance of the repository class for given $model.
+     *
+     * @param \Parable\ORM\Model $model
+     *
+     * @return Repository
+     */
+    public static function createForModel(\Parable\ORM\Model $model)
+    {
         $repository = \Parable\DI\Container::create(static::class);
-
         $repository->setModel($model);
-
         return $repository;
     }
 }