+++ tests/Components/ORM/Query/ConditionTest.php
@@ -18,7 +18,6 @@ class ConditionTest extends \Parable\Tests\Components\ORM\Base
 
         $this->query = \Parable\ORM\Query::createInstance();
         $this->query->setTableName('user');
-        $this->query->setTableKey('id');
 
         $this->condition->setQuery($this->query);
     }